refactor: cost도 롱으로 바꿔서 게산하도록 수정

+++ src/main/java/lotto/lottoService/LottoMainService.java
@@ -109,7 +109,7 @@ public class LottoMainService {
             }
             sumPrize += LottoPrize.getPrize(i, false) * stats.getHitNumberValue(i);
         }
-        return (double) sumPrize / Integer.parseInt(cost);
+        return (double) sumPrize / Long.parseLong(cost);
     }
 
 }
